requirements.txt
streamlit>=1.29.0
opencv-python-headless>=4.8.0
mediapipe>=0.10.30
numpy>=1.24.0
Pillow>=10.0.0